Added allies, individual wave spawns, turrets now fire
Blue capsule allies that serve a similar function the enemies do, but are friendly to the player and are actually how the player will win. Turrets now fire, but only at enemies. It will be easy to make enemy turrets, I just was done for the day and hadn't done that yet. Also allies have their own wave management and spawning, just like the enemies.